EDITOR-#26 move controlMap interface to slice.d.ts
